Refactored openshift plugin to make it importable as dependency
